feat(mps-sync-plugin): persist bindings to .mps/modelix.xml and restore during startup
